Fix compilation errors, add debug prints

+++ src/kmers.h
@@ -105,7 +105,7 @@ inline char NucleotideAtIndex(kmer_t encoded, int k, int index) {
 /// Return the number 0 - 3 correcponding to index-th nucleotide from the encoded k-mer.
 template <typename kmer_t>
 inline uint8_t NucleotideIndexAtIndex(kmer_t encoded, int k, int index) {
-    return (uint64_t)((encoded >> ((k - index - kmer_t(1)) << kmer_t(1))) & kmer_t(3));
+    return (uint8_t)((encoded >> ((k - index - kmer_t(1)) << kmer_t(1))) & kmer_t(3));
 }
 
 /// Convert the encoded KMer representation to string.
